send multi-line messages
line-by-line might be ugly; sending everything in a big block is more
nice, but could result in too big a request. instead of trying to split
things up or whatever, we just die() if it's more than 1K.

actually, telegram seems to be able to cope with way bigger message
size, but 1K should be enough for anybody.

Signed-off-by: knilch0r <knilch0r@users.noreply.github.com>